requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
azure-common==1.1.18
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
cycler==0.10.0
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.16.1
pandas==0.24.1
protobuf==3.6.1
ptvsd==4.2.3
pycparser==2.19
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
typed-ast==1.3.0
urllib3==1.24.1
wrapt==1.11.1
